Keep test results off the filesystem
