+++ AGENTS.md
@@ -41,8 +41,8 @@ under `crates/`.
 
 Karva runs tests through a main `karva` process and `karva-worker`
 subprocesses. The binaries do not link against each other. They communicate
-through CLI arguments and a shared cache directory, and only the worker embeds
-Python.
+through CLI arguments and loopback IPC. Only coverage data uses run-scoped
+files, and only the worker embeds Python.
 
 ## Code Review Rules
