+++ crates/karva_diagnostic/Cargo.toml
@@ -18,6 +18,7 @@ annotate-snippets = { workspace = true }
 karva_logging = { workspace = true }
 karva_python_semantic = { workspace = true }
 
+anyhow = { workspace = true }
 camino = { workspace = true }
 colored = { workspace = true }
 pyo3 = { workspace = true, optional = true }
@@ -29,6 +30,7 @@ tracing = { workspace = true }
 
 [dev-dependencies]
 insta = { workspace = true }
+tempfile = { workspace = true }
 
 [lints]
 workspace = true
