+++ crates/karva_diagnostic/src/lib.rs
@@ -1,5 +1,6 @@
 //! Test outcomes, diagnostics, tracebacks, and user-facing result reporting.
 
+mod aggregate;
 mod diagnostic;
 mod render;
 mod reporter;
@@ -7,8 +8,6 @@ mod result;
 #[cfg(feature = "traceback")]
 mod traceback;
 
-pub use diagnostic::{Annotation, Diagnostic, Severity, Span, SubDiagnostic};
-pub use render::{DiagnosticFormat, DisplayDiagnosticConfig, render_diagnostic};
 pub use reporter::{DummyReporter, Reporter, TestCaseReporter};
 pub use result::{
     CapturedTestOutput, DisplayFlakyTests, FixtureFailure, FixtureUsage, FlakyTest,
@@ -17,5 +16,8 @@ pub use result::{
     TestResultStats, TestRunResult, TestRunResultParts,
 };
 
+pub use aggregate::{AggregatedResults, WorkerResults, render_worker_results};
+pub use diagnostic::{Annotation, Diagnostic, Severity, Span, SubDiagnostic};
+pub use render::{DiagnosticFormat, DisplayDiagnosticConfig, render_diagnostic};
 #[cfg(feature = "traceback")]
 pub use traceback::{Traceback, TracebackFrame, TracebackFrameSource};
