+++ crates/karva_runner/Cargo.toml
@@ -15,6 +15,7 @@ license = { workspace = true }
 karva_cache = { workspace = true }
 karva_cli = { workspace = true }
 karva_collector = { workspace = true }
+karva_diagnostic = { workspace = true }
 karva_ipc = { workspace = true }
 karva_logging = { workspace = true }
 karva_metadata = { workspace = true }
